pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PyToNotion"
version = "0.0.2"
authors = [
    { name="vania_vaniusha", email="ivan.castaneda.nazario@gmail.com" },
]
description = "Notion API with Python"
readme = "README.md"
license = { file="LICENSE.txt" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv>=1.0.1", # Especifica la versión mínima necesaria
    "requests>=2.28",
]

[project.urls]
"Homepage" = "https://github.com/Vania-Dev/PyToNotion"
"Bug Tracker" = "https://github.com/Vania-Dev/PyToNotion/issues"

[tool.hatch.build.targets.wheel]
packages = ["PyToNotion"]
